add reset wallet
